+++ Snowflake.Data.Tests/UnitTests/SFDbParameterCollectionTest.cs
@@ -25,13 +25,13 @@ namespace Snowflake.Data.Tests
         [SFFact]
         public void TestDefaultDbParameterCollection()
         {
-            Assert.Equal(0, _parameterCollection.Count);
+            Assert.Empty(_parameterCollection);
         }
 
         [SFFact]
         public void TestDbParameterCollectionCount()
         {
-            Assert.Equal(0, _parameterCollection.Count);
+            Assert.Empty(_parameterCollection);
 
             SnowflakeDbParameter parameter = new SnowflakeDbParameter();
 
@@ -96,7 +96,7 @@ namespace Snowflake.Data.Tests
             Assert.Equal(PARAM_COUNT, _parameterCollection.Count);
 
             _parameterCollection.Clear();
-            Assert.Equal(0, _parameterCollection.Count);
+            Assert.Empty(_parameterCollection);
         }
 
         [SFTheory]
@@ -164,9 +164,8 @@ namespace Snowflake.Data.Tests
             Assert.Equal(PARAM_COUNT - 1, _parameterCollection.IndexOf(paramName));
         }
 
-        [SFTheory]
-        [MemberData(nameof(AllSFDataTypes))]
-        public void TestDbParameterCollectionIndexOfNameNotExists(SFDataType SFDataType)
+        [SFFact]
+        public void TestDbParameterCollectionIndexOfNameNotExists()
         {
             int expectedParameterIndex = -1;
             string paramName = "1";