+++ Snowflake.Data.Tests/UnitTests/Tools/FileOperationsWindowsTest.cs
@@ -20,7 +20,8 @@ namespace Snowflake.Data.Tests.UnitTests.Tools
         private static readonly string s_workingDirectory = Path.Combine(TempUtil.GetTempPath(), s_relativeWorkingDirectory);
         private static readonly string s_content = "random text";
         private static readonly string s_fileName = "testfile";
-        public static void Before()
+
+        private static void Before()
         {
             if (!Directory.Exists(s_workingDirectory))
             {
@@ -29,7 +30,8 @@ namespace Snowflake.Data.Tests.UnitTests.Tools
 
             s_fileOperations = new FileOperations();
         }
-        public static void After()
+
+        private static void After()
         {
             Directory.Delete(s_workingDirectory, true);
         }